+++ src/Ad/AbstractAdNode.php
@@ -2,8 +2,9 @@
 
 namespace OpiyOrg\Vast\Ad;
 
-use OpiyOrg\Vast\Creative\AbstractLinearCreative;
+use OpiyOrg\Vast\Creative\AbstractCreative;
 use OpiyOrg\Vast\Document\AbstractNode;
+use OpiyOrg\Vast\ElementBuilder;
 
 abstract class AbstractAdNode extends AbstractNode
 {
@@ -11,6 +12,11 @@ abstract class AbstractAdNode extends AbstractNode
      * @var \DOMElement
      */
     private $domElement;
+
+    /**
+     * @var ElementBuilder
+     */
+    protected $vastElementBuilder;
 
     /**
      * @var \DOMElement
@@ -36,12 +42,13 @@ abstract class AbstractAdNode extends AbstractNode
 
     /**
      * @param \DOMElement $adDomElement instance of \Vast\Ad element
+     * @param ElementBuilder $vastElementBuilder
      */
-    public function __construct(\DOMElement $adDomElement)
+    public function __construct(\DOMElement $adDomElement, ElementBuilder $vastElementBuilder)
     {
         $this->adDomElement = $adDomElement;
-
         $this->domElement = $this->adDomElement->getElementsByTagName($this->getType())->item(0);
+        $this->vastElementBuilder = $vastElementBuilder;
     }
 
     /**
@@ -142,13 +149,19 @@ abstract class AbstractAdNode extends AbstractNode
     }
 
     /**
-     * Build class name for creative of given type
+     * @return string[]
+     */
+    abstract protected function getAvailableCreativeTypes();
+
+    /**
+     * Build object for creative of given type
      *
      * @param string $type
+     * @param \DOMElement $creativeDomElement
      *
-     * @return string
+     * @return AbstractCreative
      */
-    abstract protected function buildCreativeClassName($type);
+    abstract protected function buildCreativeElement($type, \DOMElement $creativeDomElement);
 
     /**
      * Create "creative" object of given type
@@ -156,14 +169,14 @@ abstract class AbstractAdNode extends AbstractNode
      * @param string $type
      *
      * @throws \Exception
-     * @return AbstractLinearCreative
+     *
+     * @return AbstractCreative
      */
-    protected function buildCreative($type)
+    final protected function buildCreative($type)
     {
         // check type
-        $creativeClassName = $this->buildCreativeClassName($type);
-        if (!class_exists($creativeClassName)) {
-            throw new \Exception('Wrong creative specified: ' . var_export($creativeClassName, true));
+        if (!in_array($type, $this->getAvailableCreativeTypes())) {
+            throw new \InvalidArgumentException(sprintf('Wrong creative specified: %s', $type));
         }
 
         // get container
@@ -176,16 +189,17 @@ abstract class AbstractAdNode extends AbstractNode
             }
         }
 
-        // Creative dom element
+        // Creative dom element: <Creative></Creative>
         $creativeDomElement = $this->creativesDomElement->ownerDocument->createElement('Creative');
         $this->creativesDomElement->appendChild($creativeDomElement);
 
-        // Creative type dom element
+        // Creative type dom element: <Creative><Linear></Linear></Creative>
         $creativeTypeDomElement = $this->adDomElement->ownerDocument->createElement($type);
         $creativeDomElement->appendChild($creativeTypeDomElement);
 
         // object
-        $creative = new $creativeClassName($creativeDomElement);
+        $creative = $this->buildCreativeElement($type, $creativeDomElement);
+
         $this->creatives[] = $creative;
 
         return $creative;