+++ src/Ad/InLine.php
@@ -2,17 +2,22 @@
 
 namespace OpiyOrg\Vast\Ad;
 
-use OpiyOrg\Vast\Creative\AbstractLinearCreative;
-use OpiyOrg\Vast\Creative\InLine\Linear;
+use OpiyOrg\Vast\Creative\AbstractCreative;
+use OpiyOrg\Vast\Creative\InLine\Linear as InLineAdLinearCreative;
 
 class InLine extends AbstractAdNode
 {
+    /**
+     * @private
+     */
+    const CREATIVE_TYPE_LINEAR = 'Linear';
+
     /**
      * Set Ad title
      *
      * @param string $value
      *
-     * @return \OpiyOrg\Vast\Ad\InLine
+     * @return InLine
      */
     public function setAdTitle($value)
     {
@@ -22,21 +27,46 @@ class InLine extends AbstractAdNode
     }
 
     /**
-     * @inheritdoc
+     * @return string[]
      */
-    protected function buildCreativeClassName($type)
+    protected function getAvailableCreativeTypes()
     {
-        return '\\OpiyOrg\\Vast\\Creative\\InLine\\' . $type;
+        return array(
+            self::CREATIVE_TYPE_LINEAR,
+        );
+    }
+
+    /**
+     * @param string $type
+     * @param \DOMElement $creativeDomElement
+     *
+     * @return AbstractCreative|InLineAdLinearCreative
+     */
+    protected function buildCreativeElement($type, \DOMElement $creativeDomElement)
+    {
+        switch ($type) {
+            case self::CREATIVE_TYPE_LINEAR:
+                $creative = $this->vastElementBuilder->createInLineAdLinearCreative($creativeDomElement);
+                break;
+            default:
+                throw new \RuntimeException(sprintf('Unknown Wrapper creative type %s', $type));
+        }
+
+        return $creative;
     }
 
     /**
      * Create Linear creative
      *
      * @throws \Exception
-     * @return AbstractLinearCreative
+     *
+     * @return InLineAdLinearCreative
      */
     public function createLinearCreative()
     {
-        return $this->buildCreative('Linear');
+        /** @var InLineAdLinearCreative $creative */
+        $creative = $this->buildCreative(self::CREATIVE_TYPE_LINEAR);
+
+        return $creative;
     }
 }