+++ src/Ad/Wrapper.php
@@ -2,8 +2,16 @@
 
 namespace OpiyOrg\Vast\Ad;
 
+use OpiyOrg\Vast\Creative\AbstractCreative;
+use OpiyOrg\Vast\Creative\Wrapper\Linear as WrapperAdLinearCreative;
+
 class Wrapper extends AbstractAdNode
 {
+    /**
+     * @private
+     */
+    const CREATIVE_TYPE_LINEAR = 'Linear';
+
     /**
      * URI of ad tag of downstream Secondary Ad Server
      *
@@ -17,20 +25,44 @@ class Wrapper extends AbstractAdNode
     }
 
     /**
-     * @inheritdoc
+     * @return string[]
      */
-    protected function buildCreativeClassName($type)
+    protected function getAvailableCreativeTypes()
     {
-        return '\\OpiyOrg\\Vast\\Creative\\Wrapper\\' . $type;
+        return array(
+            self::CREATIVE_TYPE_LINEAR,
+        );
+    }
+
+    /**
+     * @param string $type
+     * @param \DOMElement $creativeDomElement
+     *
+     * @return AbstractCreative|WrapperAdLinearCreative
+     */
+    protected function buildCreativeElement($type, \DOMElement $creativeDomElement)
+    {
+        switch ($type) {
+            case self::CREATIVE_TYPE_LINEAR:
+                $creative = $this->vastElementBuilder->createWrapperAdLinearCreative($creativeDomElement);
+                break;
+            default:
+                throw new \RuntimeException(sprintf('Unknown Wrapper creative type %s', $type));
+        }
+
+        return $creative;
     }
 
     /**
      * Create Linear creative
      *
-     * @return \OpiyOrg\Vast\Creative\Wrapper\Linear
+     * @return WrapperAdLinearCreative
      */
     public function createLinearCreative()
     {
-        return $this->buildCreative('Linear');
+        /** @var WrapperAdLinearCreative $creative */
+        $creative = $this->buildCreative(self::CREATIVE_TYPE_LINEAR);
+
+        return $creative;
     }
 }