+++ src/Creative/AbstractCreative.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace OpiyOrg\Vast\Creative;
+
+use OpiyOrg\Vast\Document\AbstractNode;
+
+abstract class AbstractCreative extends AbstractNode
+{
+
+}