+++ src/Creative/AbstractLinearCreative.php
@@ -2,10 +2,21 @@
 
 namespace OpiyOrg\Vast\Creative;
 
-use OpiyOrg\Vast\Document\AbstractNode;
+use OpiyOrg\Vast\ElementBuilder;
 
-abstract class AbstractLinearCreative extends AbstractNode
+abstract class AbstractLinearCreative extends AbstractCreative
 {
+    /**
+     * @var ElementBuilder
+     */
+    protected $vastElementBuilder;
+
+    /**
+     * this event should be used to indicate when the player considers that it has loaded
+     * and buffered the creative’s media and assets either fully or to the extent that it is ready to play the media.
+     */
+    const EVENT_TYPE_LOADED = 'loaded';
+
     /**
      * not to be confused with an impression, this event indicates that an individual creative
      * portion of the ad was viewed. An impression indicates the first frame of the ad was displayed; however
@@ -90,6 +101,8 @@ abstract class AbstractLinearCreative extends AbstractNode
     const EVENT_TYPE_PROGRESS = 'progress';
 
     /**
+     * Dom Element of <Creative></Creative>
+     *
      * @var \DOMElement
      */
     private $linearCreativeDomElement;
@@ -106,13 +119,17 @@ abstract class AbstractLinearCreative extends AbstractNode
 
     /**
      * @param \DOMElement $linearCreativeDomElement
+     * @param ElementBuilder $vastElementBuilder
      */
-    public function __construct(\DOMElement $linearCreativeDomElement)
+    public function __construct(\DOMElement $linearCreativeDomElement, ElementBuilder $vastElementBuilder)
     {
         $this->linearCreativeDomElement = $linearCreativeDomElement;
+        $this->vastElementBuilder = $vastElementBuilder;
     }
 
     /**
+     * Dom Element of <Creative></Creative>
+     *
      * @return \DOMElement
      */
     protected function getDomElement()
@@ -128,6 +145,7 @@ abstract class AbstractLinearCreative extends AbstractNode
     public static function getEventList()
     {
         return array(
+            self::EVENT_TYPE_LOADED,
             self::EVENT_TYPE_CREATIVEVIEW,
             self::EVENT_TYPE_START,
             self::EVENT_TYPE_FIRSTQUARTILE,
@@ -168,7 +186,10 @@ abstract class AbstractLinearCreative extends AbstractNode
         }
         
         $this->videoClicksDomElement = $this->linearCreativeDomElement->ownerDocument->createElement('VideoClicks');
-        $this->linearCreativeDomElement->firstChild->appendChild($this->videoClicksDomElement);
+        $this->linearCreativeDomElement
+            ->getElementsByTagName('Linear')
+            ->item(0)
+            ->appendChild($this->videoClicksDomElement);
         
         return $this->videoClicksDomElement;
     }
@@ -263,7 +284,10 @@ abstract class AbstractLinearCreative extends AbstractNode
             ->ownerDocument
             ->createElement('TrackingEvents');
 
-        $this->linearCreativeDomElement->firstChild->appendChild($this->trackingEventsDomElement);
+        $this->linearCreativeDomElement
+            ->getElementsByTagName('Linear')
+            ->item(0)
+            ->appendChild($this->trackingEventsDomElement);
         
         return $this->trackingEventsDomElement;
     }
@@ -287,14 +311,69 @@ abstract class AbstractLinearCreative extends AbstractNode
         
         // add event attribute
         $trackingDomElement->setAttribute('event', $event);
-        
+
         // create cdata
         $cdata = $this->linearCreativeDomElement->ownerDocument->createCDATASection($url);
         $trackingDomElement->appendChild($cdata);
-        
+
         return $this;
     }
 
+    /**
+     * @param string $url
+     * @param int|string $offset seconds or time in format "H:m:i" or percents in format "n%"
+     *
+     * @return $this
+     */
+    public function addProgressTrackingEvent($url, $offset)
+    {
+        // create Tracking
+        $trackingDomElement = $this->linearCreativeDomElement->ownerDocument->createElement('Tracking');
+        $this->getTrackingEventsDomElement()->appendChild($trackingDomElement);
+
+        // add event attribute
+        $trackingDomElement->setAttribute('event', self::EVENT_TYPE_PROGRESS);
+
+        // add offset attribute
+        if (is_numeric($offset)) {
+            $offset = $this->secondsToString($offset);
+        }
+        $trackingDomElement->setAttribute('offset', $offset);
+
+        // create cdata
+        $cdata = $this->linearCreativeDomElement->ownerDocument->createCDATASection($url);
+        $trackingDomElement->appendChild($cdata);
+
+        return $this;
+    }
+
+    /**
+     * Convert seconds to H:m:i
+     * Hours could be more than 24
+     *
+     * @param mixed $seconds
+     *
+     * @return string
+     */
+    protected function secondsToString($seconds)
+    {
+        $seconds = (int) $seconds;
+
+        $time = array();
+
+        // get hours
+        $hours = floor($seconds / 3600);
+        $time[] = str_pad($hours, 2, '0', STR_PAD_LEFT);
+
+        // get minutes
+        $seconds = $seconds % 3600;
+        $time[] = str_pad(floor($seconds / 60), 2, '0', STR_PAD_LEFT);
+
+        // get seconds
+        $time[] = str_pad($seconds % 60, 2, '0', STR_PAD_LEFT);
+
+        return implode(':', $time);
+    }
 
     /**
      * Get id
@@ -334,6 +413,5 @@ abstract class AbstractLinearCreative extends AbstractNode
 
         return $this;
     }
-
 
 }