+++ src/Creative/InLine/Linear.php
@@ -4,6 +4,7 @@ namespace OpiyOrg\Vast\Creative\InLine;
 
 use OpiyOrg\Vast\Creative\AbstractLinearCreative;
 use OpiyOrg\Vast\Creative\InLine\Linear\MediaFile;
+use OpiyOrg\Vast\Creative\InLine\Linear\AdParameters;
 
 class Linear extends AbstractLinearCreative
 {
@@ -11,6 +12,11 @@ class Linear extends AbstractLinearCreative
      * @var \DOMElement
      */
     private $mediaFilesDomElement;
+
+    /**
+     * @var \DOMElement
+     */
+    private $adParametersDomElement;
 
     /**
      * Set duration value
@@ -25,7 +31,7 @@ class Linear extends AbstractLinearCreative
         $durationDomElement = $this->getDomElement()->getElementsByTagName('Duration')->item(0);
         if (!$durationDomElement) {
             $durationDomElement = $this->getDomElement()->ownerDocument->createElement('Duration');
-            $this->getDomElement()->firstChild->appendChild($durationDomElement);
+            $this->getDomElement()->getElementsByTagName('Linear')->item(0)->appendChild($durationDomElement);
         }
 
         // set value
@@ -48,7 +54,10 @@ class Linear extends AbstractLinearCreative
             $this->mediaFilesDomElement = $this->getDomElement()->getElementsByTagName('MediaFiles')->item(0);
             if (!$this->mediaFilesDomElement) {
                 $this->mediaFilesDomElement = $this->getDomElement()->ownerDocument->createElement('MediaFiles');
-                $this->getDomElement()->firstChild->appendChild($this->mediaFilesDomElement);
+                $this->getDomElement()
+                    ->getElementsByTagName('Linear')
+                    ->item(0)
+                    ->appendChild($this->mediaFilesDomElement);
             }
         }
 
@@ -57,35 +66,37 @@ class Linear extends AbstractLinearCreative
         $this->mediaFilesDomElement->appendChild($mediaFileDomElement);
 
         // object
-        return new MediaFile($mediaFileDomElement);
+        return $this->vastElementBuilder->createInLineAdLinearCreativeMediaFile($mediaFileDomElement);
     }
 
     /**
-     * Convert seconds to H:m:i
-     * Hours could be more than 24
+     * @param array|string $params
      *
-     * @param mixed $seconds
-     *
-     * @return string
+     * @return self
      */
-    private function secondsToString($seconds)
+    public function setAdParameters($params)
     {
-        $seconds = (int) $seconds;
+        $this->adParametersDomElement = $this->getDomElement()->getElementsByTagName('AdParameters')->item(0);
+        if (!$this->adParametersDomElement) {
+            $this->adParametersDomElement = $this->getDomElement()->ownerDocument->createElement('AdParameters');
+            $this->getDomElement()->getElementsByTagName('Linear')->item(0)->appendChild($this->adParametersDomElement);
+        }
 
-        $time = array();
+        if (is_array($params)) {
+            $params = json_encode($params);
+        }
 
-        // get hours
-        $hours = floor($seconds / 3600);
-        $time[] = str_pad($hours, 2, '0', STR_PAD_LEFT);
+        $cdata = $this->adParametersDomElement->ownerDocument->createCDATASection($params);
 
-        // get minutes
-        $seconds = $seconds % 3600;
-        $time[] = str_pad(floor($seconds / 60), 2, '0', STR_PAD_LEFT);
+        // update CData
+        if ($this->adParametersDomElement->hasChildNodes()) {
+            $this->adParametersDomElement->replaceChild($cdata, $this->adParametersDomElement->firstChild);
+        } // insert CData
+        else {
+            $this->adParametersDomElement->appendChild($cdata);
+        }
 
-        // get seconds
-        $time[] = str_pad($seconds % 60, 2, '0', STR_PAD_LEFT);
-
-        return implode(':', $time);
+        return $this;
     }
 
     /**
@@ -98,7 +109,27 @@ class Linear extends AbstractLinearCreative
             $time = $this->secondsToString($time);
         }
 
-        $this->getDomElement()->firstChild->setAttribute('skipoffset', $time);
+        $this->getDomElement()->getElementsByTagName('Linear')->item(0)->setAttribute('skipoffset', $time);
+
+        return $this;
+    }
+
+
+    /**
+     * <UniversalAdId> required element for the purpose of tracking ad creative, he added in VAST 4.0 spec.
+     * Paragraph 3.7.1
+     * https://iabtechlab.com/wp-content/uploads/2018/11/VAST4.1-final-Nov-8-2018.pdf
+     *
+     * @param int|string $idRegistry
+     * @param int|string $universalAdId
+     * @return $this
+     */
+    public function setUniversalAdId($idRegistry, $universalAdId)
+    {
+        $universalAdIdDomElement = $this->getDomElement()->ownerDocument->createElement('UniversalAdId');
+        $universalAdIdDomElement->nodeValue = $universalAdId;
+        $universalAdIdDomElement->setAttribute("idRegistry", $idRegistry);
+        $this->getDomElement()->insertBefore($universalAdIdDomElement, $this->getDomElement()->firstChild);
 
         return $this;
     }